+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.38](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.37...@kong-ui-public/analytics-metric-provider@1.1.38) (2023-09-20)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [1.1.37](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.36...@kong-ui-public/analytics-metric-provider@1.1.37) (2023-09-18)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [1.1.36](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.35...@kong-ui-public/analytics-metric-provider@1.1.36) (2023-09-16)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [1.1.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.34...@kong-ui-public/analytics-metric-provider@1.1.35) (2023-09-14)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [1.1.34](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.33...@kong-ui-public/analytics-metric-provider@1.1.34) (2023-09-14)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [1.1.33](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@1.1.32...@kong-ui-public/analytics-metric-provider@1.1.33) (2023-09-13)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider