+++ packages/core/misc-widgets/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.38](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/misc-widgets@0.5.37...@kong-ui-public/misc-widgets@0.5.38) (2023-09-20)
+
+**Note:** Version bump only for package @kong-ui-public/misc-widgets
+
+
+
+
+
+## [0.5.37](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/misc-widgets@0.5.36...@kong-ui-public/misc-widgets@0.5.37) (2023-09-18)
+
+**Note:** Version bump only for package @kong-ui-public/misc-widgets
+
+
+
+
+
+## [0.5.36](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/misc-widgets@0.5.35...@kong-ui-public/misc-widgets@0.5.36) (2023-09-16)
+
+**Note:** Version bump only for package @kong-ui-public/misc-widgets
+
+
+
+
+
+## [0.5.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/misc-widgets@0.5.34...@kong-ui-public/misc-widgets@0.5.35) (2023-09-14)
+
+**Note:** Version bump only for package @kong-ui-public/misc-widgets
+
+
+
+
+
 ## [0.5.34](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/misc-widgets@0.5.33...@kong-ui-public/misc-widgets@0.5.34) (2023-09-13)
 
 **Note:** Version bump only for package @kong-ui-public/misc-widgets