+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@1.3.5...@kong-ui-public/entities-shared@1.3.6) (2023-09-20)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [1.3.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@1.3.4...@kong-ui-public/entities-shared@1.3.5) (2023-09-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [1.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@1.3.3...@kong-ui-public/entities-shared@1.3.4) (2023-09-16)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [1.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@1.3.2...@kong-ui-public/entities-shared@1.3.3) (2023-09-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
 ## [1.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@1.3.1...@kong-ui-public/entities-shared@1.3.2) (2023-09-13)
 
 **Note:** Version bump only for package @kong-ui-public/entities-shared